# PERF: short-circuit allocations in infer_dtype, ensure_datetime64ns