```
from pandas._libs.tslibs.conversion import *

arr = np.arange(10**6).view("M8[ns]")

%timeit ensure_datetime64ns(arr, copy=False)
423 ns ± 10.3 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)  # <- PR
3.5 µs ± 17.3 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)  # <- master

%timeit ensure_datetime64ns(arr, copy=True)
517 µs ± 11.4 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)   # <- PR
1.77 ms ± 4.44 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)  # <- master


from pandas._libs.lib import *

farr = np.arange(10**6)[::2]

%timeit infer_dtype(farr, skipna=True)
2.94 µs ± 48.2 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)  # <- PR
412 µs ± 34.4 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)  # <- master
```