@jbrockmendel 2021-12-20
> looks legit (32-bit build)

those show up in the logs pretty frequently but don't cause failures.  i'll wait until the other CI issues are resolved to re-push and see if its an actual problem here